Further improvement to JWT token signing
