+++ blueocean-jwt/src/main/java/io/jenkins/blueocean/auth/jwt/JwtToken.java
@@ -3,6 +3,7 @@ package io.jenkins.blueocean.auth.jwt;
 import hudson.ExtensionList;
 import hudson.ExtensionPoint;
 import net.sf.json.JSONObject;
+import org.jose4j.jwx.HeaderParameterNames;
 import org.kohsuke.stapler.HttpResponse;
 import org.kohsuke.stapler.StaplerRequest;
 import org.kohsuke.stapler.StaplerResponse;
@@ -11,6 +12,7 @@ import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
 import javax.servlet.ServletException;
 import java.io.IOException;
+import java.util.UUID;
 
 /**
  * Generates JWT token
@@ -41,7 +43,28 @@ public abstract class JwtToken implements HttpResponse, ExtensionPoint{
      *
      * @return base64 representation of JWT token
      */
-    public abstract String sign();
+    public String sign(){
+        for(JwtTokenDecorator decorator: JwtTokenDecorator.all()){
+            decorator.decorate(this);
+        }
+
+        /**
+         *  kid might have been set already by using {@link #header} or {@link JwtTokenDecorator}, if present use it
+         *  otherwise use the default kid
+         */
+        String keyId = (String)header.get(HeaderParameterNames.KEY_ID);
+        if(keyId == null){
+            keyId = UUID.randomUUID().toString().replace("-", "");
+        }
+        return sign(keyId);
+    }
+
+    /**
+     * Do the actual signing using this keyId
+     * @param keyId keyId
+     * @return signed value as String
+     */
+    public abstract String sign(String keyId);
 
     @Override
     public void generateResponse(StaplerRequest req, StaplerResponse rsp, Object node) throws IOException, ServletException {
